Moved the azure migration service to dgconn, since the other is public and the users can enter. It can be accessed with the url: http://localhost:8080/dashboard/api/azure/blob Added some security (checks the user is logged and is admin, so no anonymous users can launch it) Moved LegalFiles repository to persistence project, because its common with several modules

+++ wifi4eu-portal/wifi4eu-portal-service/src/main/java/wifi4eu/wifi4eu/service/user/UserService.java
@@ -675,8 +675,8 @@ public class UserService {
     public boolean checkIfVoucherAwarded(UserDTO userDTO) {
         List<Integer> voucherAwarded = userRepository.getIfUserHasVouchersAwarded(userDTO.getId());
         if (Validator.isNotNull(voucherAwarded)) {
-            for (Integer vouch : voucherAwarded) {
-                if (vouch.intValue() == 1) {
+            for (int vouch : voucherAwarded) {
+                if (vouch == 1) {
                     return true;
                 }
             }
